+++ src/library/predicates/index.ts
@@ -1,3 +1,3 @@
 export * from './Vault';
-export { defaultValues } from './helpers';
+export * from './helpers';
 export * from './types';
